add .vscode to .gitignore and create bike.js
