+++ .gitignore
@@ -2,3 +2,4 @@ node_modules
 npm-debug.log
 /coverage
 .idea
+.vscode
