+++ lib/item.js
@@ -6,6 +6,8 @@ const iconv = require('iconv-lite');
 
 const cleanString = require('./utils').cleanString;
 
+const Bike = require('./bike.js');
+
 const Item = function (options) {
     if (!(this instanceof Item)) 
         return new Item(options);
@@ -170,35 +172,9 @@ Item.prototype.getDetails = function () {
     });
 };
 
-/* 
- * Create ad.bike.cubic_capacity, ad.bike.mileage and ad.bike.regdate
- * Make them with getters which get requested value finding it into object
- * returned by getDetails.
- */
-
-Bike = function(ad) {
-    // store reference to ad.
-    this.ad = ad;
-}
-
 Object.defineProperty(Item.prototype, 'bike', { get : function() {
-    return new Bike(this);
+    return new Bike.Bike(this);
 }});
-
-/**
- * getter return '' if requested value is not defined.
- */
-function _makeGetter(name) {
-    Object.defineProperty(Bike.prototype, name, { get : function() {
-        var attr1 = this.ad.details.attributes.find(attr => attr.key == name);
-        var v = attr1 ? attr1.value : '';
-        return v;
-    }});
-}
-
-_makeGetter('cubic_capacity');
-_makeGetter('mileage');
-_makeGetter('regdate');
 
 Item.prototype.getPhoneNumber = function () {
     var self = this;
